chore: move auth -> @lib
