+++ src/@lib/index.ts
@@ -1,5 +1,6 @@
 export * from "./@domain";
 export * from "./@types";
+export * from "./auth";
 export * from "./cmd";
 export * from "./errors";
 export * from "./logger";
